Issue #7440: Split gulp tasks into multiple files
