+++ package.json
@@ -42,6 +42,7 @@
     "gulp-cache": "^0.4.6",
     "gulp-clean-css": "^3.9.4",
     "gulp-concat": "^2.6.0",
+    "gulp-hub": "^4.2.0",
     "gulp-imagemin": "^5.0.3",
     "gulp-jshint": "^2.1.0",
     "gulp-notify": "^2.2.0",
@@ -62,8 +63,6 @@
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-gravatar": "^2.6.3",
-    "run-sequence": "^1.2.2",
-    "webpack": "^3.12.0",
-    "yargs": "^11.1.0"
+    "webpack": "^3.12.0"
   }
 }
